+++ libraries/AGS3871/CHANGELOG.md
@@ -6,5 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2025-01-04
+- fix + add examples
+- rewrite **AGS3871_get_registers.ino**
+- add **lastStatus()**
+- change return type **bool dataReady()**
+- update readme.md
+
+
 ## [0.1.0] - 2025-01-01
 - initial version (based on AGS02MA)